started Rating recipe functionality. added GET endpoint for rating User left on Recipe.

+++ Models/Rating.cs
@@ -7,10 +7,10 @@ namespace foodtopia.Models
         public Guid Id { get; set; }
         public Guid RecipeId { get; set; }
         public Guid UserId { get; set; }
-        [Range(1, 5, ErrorMessage = "Taste rating must be between 1 and 5")]
-        public decimal TasteRating { get; set; }
-        [Range(1, 5, ErrorMessage = "Difficulty rating must be between 1 and 5")]
-        public decimal DifficultyRating { get; set; }
+        [Range(1.0, 5.0, ErrorMessage = "Taste rating must be between 1 and 5")]
+        public double TasteRating { get; set; }
+        [Range(1.0, 5.0, ErrorMessage = "Difficulty rating must be between 1 and 5")]
+        public double DifficultyRating { get; set; }
         public DateTime ReviewedAt { get; set; } = DateTime.UtcNow;
 
         public required Recipe Recipe { get; set; }
